Rename the io module to surfer.py
This avoids clashes with the Python standard library io module.
