+++ src/verde/__init__.py
@@ -21,7 +21,6 @@ from .coordinates import (
     scatter_points,
 )
 from .distances import median_distance
-from .io import load_surfer
 from .mask import convexhull_mask, distance_mask
 from .model_selection import (
     BlockKFold,
@@ -33,6 +32,7 @@ from .neighbors import KNeighbors
 from .projections import project_grid, project_region
 from .scipygridder import Cubic, Linear
 from .spline import Spline, SplineCV
+from .surfer import load_surfer
 from .trend import Trend
 from .utils import grid_to_table, make_xarray_grid, maxabs, variance_to_weights
 from .vector import Vector, VectorSpline2D
